backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
email-validator==2.1.0
python-dotenv==1.0.0
